Convert start sytem from shortcut to pyw
I had just created a shortcut system to start savitime.
But it started the IDLE also.  I also tried double-clicking
on the python script that the start shortcut called.  That
worked but it also started a cmd window.

Moving the start script out of the subdir and giving it
the extensio ".pyw" instead of ".py" made it totally work.